Complete Prey.hpp & Prey.cpp

+++ code/Agent.cpp
@@ -7,7 +7,7 @@ Agent::Agent(mt19937_64 &mt, const Chromosome &_chr)
 
     VisionAngle = 2 * PI - (N_SEGMENT_SENSOR * EACH_SEGMENT_ANGLE) / 180 * PI;
 }
-Agent::Agent(mt19937_64 &mt, const Chromosome &_chr, const int &_id)
+Agent::Agent(mt19937_64 &mt, const Chromosome &_chr, const long long &_id)
 {
     F_live = true;
